pyproject.toml
[tool.poetry]
name = "swarmos"
version = "0.1.0"
description = "Local-first multi-agent control plane with hybrid debate mechanism"
authors = ["SwarmOS Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
qdrant-client = "^1.11.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
litellm = "^1.52.0"
httpx = "^0.27.2"
jinja2 = "^3.1.4"
python-multipart = "^0.0.12"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
openai = "^1.54.0"
anthropic = "^0.34.0"
google-generativeai = "^0.8.0"
tiktoken = "^0.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]

